requirements/base.txt
django==1.10.5
MySQL-python==1.2.5
djangorestframework==3.5.3
django-rest-swagger==2.1.1
django-cors-headers==2.0.0
django-celery==3.1.17
celery[redis]<4